Splits the public interface from the internals.
